+++ test/test_ecc_throws.jl
@@ -5,3 +5,6 @@ using QuantumClifford.ECC: ReedMuller
 
 @test_throws ArgumentError ReedMuller(-1, 3)
 @test_throws ArgumentError ReedMuller(1, 0)
+
+@test_throws ArgumentError ReedSolomon(-1, -1)
+@test_throws ArgumentError ReedSolomon(31, 35)